Move IRIBuild to Build
This necessitates a lot of other changes. Build generates all named
entities, which breaks all of the tests. All methods on Ontology which
created something and added it to the ontology are gone. New methods,
insert and declare have been added.

All the binaries are disabled and reader.rs and writer.rs are excluded
as they are not going to work for a while.