Fix ontology annotation handling
Previously ontology annotation handling was only working where,
paradoxically, the annotation property was not defined or recognised
as an annotation property. The unit test appears to have been
generated using an old version of the OWL which was missing this
declaration.

This has now been updated, as have the unit tests which have been
expanded and buboized.