Ensure test file generation is consistent
Previously, test files were generated in a variety of different ways, both manually and using bubo; it was often unclear why files were generated like this. Where it is possible to do, test files are now generated using bubo consistently.

An upper level make file has been added for convenience, and a README file which documents the generation of these files.

DataUnionOf constructs are now supported, and the erroneous test file that was hiding this breakage has been updated.