Update error handling in the RDF parser
The semantics of the error handling method was broken, apparently
happening in two different places. The parse method reported that it
could return an Error when it could not. The "as" and "ref" methods
both were capable of returning Errors.

The code has now been updated so that parse really can return an
error, rather than delaying the reporting of this error to some point
in the future, while the "as" and "ref" methods now cannot return an
error. Documentation has been updated to note that the ontology that
is returned might be empty; this was true before, but has now been
clarified.

Documentation and naming in DeclarationMappedIndex has been updated
for clarity.

Add a failing, ignored test for punning bug, which is the motivation
for all other changes.